meshed rpi code with authorization code.
