+++ index.js
@@ -17,9 +17,9 @@ const resourceRouter = require('./router/resource-router.js');
 const bodyParser = require('body-parser');
 const mongoose = require('mongoose');
 const requestIp = require('request-ip');
+const socket = require('socket.io');
 const cors = require('cors');
 
-const app = express();
 mongoose.connect(process.env.MONGODB_URI);
 app.use(cors());
 app.use(bodyParser.urlencoded({extended:true}));
@@ -37,6 +37,19 @@ app.use((request, response) => {
   response.sendFile(__dirname + '/public/signin.html');
 });
 
-app.listen(PORT, () => {
+const server = app.listen(PORT, () => {
   console.log(`http://localhost:${PORT}`);
+});
+
+const io = socket(server);
+
+io.on('connection', function(socket){
+  let clientIp = [];
+  io.emit('ledStatus', isLedOn);
+  clientIp.push(socket.request.connection.remoteAddress);
+  console.log('made socket connection', socket.id, clientIp);
+
+  socket.on('ledStatus', (isLedOn) => {
+    io.sockets.emit('ledStatus', isLedOn);
+  });
 });
