+++ lib/basic-auth-middleware.js
