+++ lib/bearer-auth-middlewear.js
