+++ scripts/save-user-with-resource.js
